+++ test/eventmanager_spec.js
@@ -12,11 +12,11 @@ describe('EventManager', function() {
     var err = null;
     try {
       em = new EventManager();
-    } catch(e) {
+    } catch (e) {
       err = e;
-    };
+    }
     expect(err).toEqual(Error('use the EventManager.getInstance() method'));
-  })
+  });
 
   it('.getInstance() returns always the same instance', function() {
     var a = EventManager.getInstance();
@@ -37,17 +37,19 @@ describe('EventManager', function() {
     afterEach(function() {
       em = null;
     });
-    it('must be invoked with params: `eventOrEvents`, `emitFn`, `context`', function() {
-      var err = null;
-      expect(em instanceof EventManager).toBeTruthy();
-      expect(em.register).not.toHaveBeenCalled();
-      try {
-        em.register();
-      } catch (e) {
-        err = e;
-      }
-      expect(err).toMatch('You must pass eventName a emitFn to registerEvent()');
-    });
+    it('must be invoked with params: `eventOrEvents`, `emitFn`, `context`',
+      function() {
+        var err = null;
+        expect(em instanceof EventManager).toBeTruthy();
+        expect(em.register).not.toHaveBeenCalled();
+        try {
+          em.register();
+        } catch (e) {
+          err = e;
+        }
+        expect(err)
+          .toMatch('You must pass eventName a emitFn to registerEvent()');
+      });
     it('delegates a single event input to registerEvent()', function() {
       var eventName = 'first:event';
       var eventFn = function() {
@@ -62,9 +64,13 @@ describe('EventManager', function() {
 
     it('delegates a eventsHash to registerEvents()', function() {
       var eventsHash = {
-        'first:event' : function() {return 'i am the first event';},
-        'second:event': function() {return 'i am the second event';} 
-      }; 
+        'first:event': function() {
+          return 'i am the first event';
+        },
+        'second:event': function() {
+          return 'i am the second event';
+        }
+      };
 
       em.register(eventsHash);
       expect(em.register).toHaveBeenCalled();
@@ -75,52 +81,52 @@ describe('EventManager', function() {
     });
   });
 
-describe('emit() method', function() {
-  beforeEach(function() {
-    em = EventManager.getInstance()
-    em.registeredSubs = {};
-    spyOn(em, 'emit').and.callThrough();
-    spyOn(em, 'register').and.callThrough();
-  });    
+  describe('emit() method', function() {
+    beforeEach(function() {
+      em = EventManager.getInstance();
+      em.registeredSubs = {};
+      spyOn(em, 'emit').and.callThrough();
+      spyOn(em, 'register').and.callThrough();
+    });
 
-  afterEach(function() {
-    em = null;
-  });
+    afterEach(function() {
+      em = null;
+    });
 
-  it('is defined and callable', function() {
-    em.emit('first:event', {a:1}, this);
-    expect(em.emit).toHaveBeenCalled();
-  });
-  it('emits an event defined in registeredSubs', function() {
-    var eventName = 'first:event';
-    var eventFn = function() {
-      console.log('first:event');
-    };
-    em.register(eventName, eventFn);
-    spyOn(em.registeredSubs[eventName][0], 'emitFn').and.callThrough();
-    em.emit(eventName);
-    expect(em.registeredSubs[eventName][0].emitFn).toHaveBeenCalled();
-  });
-  it('emits all emitFunctions registered under one event', function() {
-    var counter = 0;
-    var eventName = 'first:event';
-    var firstEmitFn = function() {
-      return counter += 1;
-    };
-    var secondEmitFn = function() {
-      return counter += 1;
-    }
+    it('is defined and callable', function() {
+      em.emit('first:event', {a: 1}, this);
+      expect(em.emit).toHaveBeenCalled();
+    });
+    it('emits an event defined in registeredSubs', function() {
+      var eventName = 'first:event';
+      var eventFn = function() {
+        console.log('first:event');
+      };
+      em.register(eventName, eventFn);
+      spyOn(em.registeredSubs[eventName][0], 'emitFn').and.callThrough();
+      em.emit(eventName);
+      expect(em.registeredSubs[eventName][0].emitFn).toHaveBeenCalled();
+    });
+    it('emits all emitFunctions registered under one event', function() {
+      var counter = 0;
+      var eventName = 'first:event';
+      var firstEmitFn = function() {
+        return counter += 1;
+      };
+      var secondEmitFn = function() {
+        return counter += 1;
+      };
 
-    em.register(eventName, firstEmitFn);
-    expect(em.register).toHaveBeenCalled();
-    em.register(eventName, secondEmitFn);
-    expect(em.register).toHaveBeenCalled();
-    spyOn(em.registeredSubs[eventName][0], 'emitFn').and.callThrough();
-    spyOn(em.registeredSubs[eventName][1], 'emitFn').and.callThrough();
-    em.emit(eventName);
-    expect(counter).toEqual(2);
-    expect(em.registeredSubs[eventName][0].emitFn).toHaveBeenCalled();
-    expect(em.registeredSubs[eventName][1].emitFn).toHaveBeenCalled();
+      em.register(eventName, firstEmitFn);
+      expect(em.register).toHaveBeenCalled();
+      em.register(eventName, secondEmitFn);
+      expect(em.register).toHaveBeenCalled();
+      spyOn(em.registeredSubs[eventName][0], 'emitFn').and.callThrough();
+      spyOn(em.registeredSubs[eventName][1], 'emitFn').and.callThrough();
+      em.emit(eventName);
+      expect(counter).toEqual(2);
+      expect(em.registeredSubs[eventName][0].emitFn).toHaveBeenCalled();
+      expect(em.registeredSubs[eventName][1].emitFn).toHaveBeenCalled();
+    });
   });
-});
 });